refactor(core): rename application layer
